% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Discovr.R
\name{movie_disc}
\alias{movie_disc}
\title{Discover Movies Based on your Preferences}
\usage{
movie_disc(genre = NA, rating = 0, director = NA, actor = NA)
}
\value{
movie_title
}
\description{
Discover Movies Based on your Preferences
}
\examples{
movie_disc("Action", 7, "Ridley Scott", "Mark Addy" )
movie_disc("Action", 5, ,"Tom Cruise")
}
