% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moviefinder.R
\docType{package}
\name{moviefinder-package}
\alias{moviefinder}
\alias{moviefinder-package}
\title{The Movie Finder function}
\description{
This package has 2 functions, Discovr and search.
}
\details{
The Discovr package helps you to find movies that you wish to watch based on 4 inputs.
These 4 inputs are genre, minimum imdb rating, main actor and director.
The function is called by movie_disc()

The second function is search.
This function will allow you to see all the important information of a movie
The function is called using MovieFinder
}
